savegame: middle of refactoring
