+++ test/fixtures/cookbooks/php_runtime_test/recipes/source.rb
@@ -1,0 +1,7 @@
+# comments!
+
+php_runtime 'default' do
+  version node['php']['version']
+  provider Chef::Provider::PhpRuntime::Source
+  action :install
+end
